feat: pick&place weld grasp, RL bindings, ROS gripper/TF, mm_mobile wheel fix
Physics:
- Add prismatic and fixed (weld) joints to rne_physics + Rapier backend
- Make backend ECS iteration deterministic (body/joint sync + transform
  write-back sorted by entity), fixing the flaky shoulder-motor test

Manipulation (Phase C):
- Contact-triggered weld grasp in MobileManipulatorSim (is_grasping/grasped_object):
  closing the gripper onto a body welds it to the end-effector, opening releases it
- MobileManipulatorTask::Place + MobileManipulatorEpisodeConfig::place()
- Example 26 pick-and-place smoke (grasp -> carry -> release -> settle)
- Recalibrate example 21 reach target to the now-deterministic reachable pose

RL / Python:
- rne_py MobileManipulator bindings (sim + episode, place/transport/inspect)
- Example 27 gymnasium-style MobileManipulatorPlaceEnv (degrades without gymnasium/numpy)

ROS 2:
- /gripper_command (std_msgs/Float64) and ee_link TF in mobile_manipulator mode
- Fix rne_ros2_node build: sensor_msgs/Image.is_bigendian bool -> u8
- Smoke checks peak |base_x| (mm_mobile path curves) + gripper/ee_link

Assets:
- Fix mm_mobile drive wheels: were stacked vertically (only one touched ground,
  base spun in place); relocated to a proper left/right diff-drive layout so the
  base drives forward

Verified: cargo run -p xtask -- ci and cargo run -p xtask -- ci-ros2 both pass.

+++ adapters/ros2/rne_ros2_node/src/convert.rs
@@ -68,7 +68,7 @@ pub fn to_image_message(image: &RosImage) -> sensor_msgs::msg::Image {
         height: image.height,
         width: image.width,
         encoding: image.encoding.clone(),
-        is_bigendian: image.is_bigendian,
+        is_bigendian: image.is_bigendian as u8,
         step: image.step,
         data: image.data.clone(),
     }
